# doc/api.yaml
openapi: 3.0.3
info:
  title: WASA-Exam
  description: API for the WASA Exam project.
  version: 1.0.0
security: []
...
paths:
...
  /conversations/{conversationId}/messages/{messageId}/reaction:
    parameters:
      - name: conversationId
        in: path
        required: true
        description: The ID of the conversation
        schema:
          type: string
          description: UUID of conversation
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
      - name: messageId
        in: path
        required: true
        description: The ID of the message
        schema:
          type: string
          description: UUID of message
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
    put:
      tags: [Reaction]
      summary: Add a reaction to a message
      description: Add a reaction (emoji) to a message.
      operationId: commentMessage
      security:
        - BearerAuth: []
      requestBody:
        description: Reaction details
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Reaction payload
              required: [type]
              properties:
                type:
                  type: string
                  description: Type of reaction to add
                  enum: [like, heart, laugh, sad_face, angry_face]
      responses:
        "200":
          description: Reaction added successfully
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  message:
                    type: string
                    description: Success message
                    example: "Reaction added"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                description: Error response
                properties:
                  error:
                    type: string
                    description: Error message
                    example: "Invalid reaction"
        "404":
          description: Message or conversation not found
          content:
            application/json:
              schema:
                type: object
                description: Error response
                properties:
                  error:
                    type: string
                    description: Error message
                    example: "Message not found"

    delete:
      tags: [Reaction]
      summary: Remove a reaction from a message
      description: Remove a reaction from a message.
      operationId: uncommentMessage
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          required: true
          description: The type of reaction to remove
          schema:
            type: string
            description: Reaction type
            enum: [like, heart, laugh, sad_face, angry_face]
      responses:
        "200":
          description: Reaction removed successfully
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  message:
                    type: string
                    description: Success message
                    example: "Reaction removed"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                description: Error response
                properties:
                  error:
                    type: string
                    description: Error message
        "404":
          description: Message or conversation not found
  /groups:
    post:
      tags: [Group]
      summary: Create a group
      description: Create a new group conversation.
      operationId: createGroup
      security:
        - BearerAuth: []
      requestBody:
        description: Group details
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Group creation payload
              required: [groupName, members]
              properties:
                groupName:
                  type: string
                  description: Name of the group
                  minLength: 1
                  maxLength: 100
                  pattern: "^[a-zA-Z0-9 _\\-]+$"
                members:
                  type: array
                  description: List of initial members (user IDs)
                  minItems: 1
                  maxItems: 100
                  items:
                    type: string
                    description: User UUID
                    pattern: "^[0-9a-fA-F\\-]+$"
                groupPhoto:
                  type: string
                  description: Optional group photo
                  format: byte
                  minLength: 0
                  maxLength: 5242880
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /groups/{groupId}:
    parameters:
      - name: groupId
        in: path
        required: true
        description: The ID of the group
        schema:
          type: string
          description: UUID of group
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
    post:
      tags: [Group]
      summary: Add user to group
      description: Add a user to an existing group.
      operationId: addToGroup
      security:
        - BearerAuth: []
      requestBody:
        description: User to add
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Add user payload
              required: [userId]
              properties:
                userId:
                  type: string
                  description: User UUID
                  pattern: "^[0-9a-fA-F\\-]+$"
                  minLength: 36
                  maxLength: 36
      responses:
        "204":
          description: User added

    delete:
      tags: [Group]
      summary: Remove user from group (or leave if self)
      description: Remove a user from the group.
      operationId: leaveGroup
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: query
          required: true
          description: ID of user to remove
          schema:
            type: string
            description: User UUID
            pattern: "^[0-9a-fA-F\\-]+$"
            minLength: 36
            maxLength: 36
      responses:
        "204":
          description: User removed
  /groups/{groupId}/name:
    parameters:
      - name: groupId
        in: path
        required: true
        description: The ID of the group
        schema:
          type: string
          description: UUID of group
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
    put:
      tags:
        - Group
      summary: Change group name
      description: Update the name of a group.
      operationId: setGroupName
      security:
        - BearerAuth: []
      requestBody:
        description: New group name
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Name payload
              required:
                - name
              properties:
                name:
                  type: string
                  description: New name
                  minLength: 1
                  maxLength: 100
                  pattern: "^[a-zA-Z0-9 _\\-]+$"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  id:
                    type: string
                    description: Group ID
                  name:
                    type: string
                    description: New Name
                  isGroup:
                    type: boolean
                    description: Is group flag
  /groups/{groupId}/photo:
    parameters:
      - name: groupId
        in: path
        required: true
        description: The ID of the group
        schema:
          type: string
          description: UUID of group
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
    put:
      tags:
        - Group
      summary: Set group photo
      description: Update the photo of a group.
      operationId: setGroupPhoto
      security:
        - BearerAuth: []
      requestBody:
        description: New group photo
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Photo payload
              required:
                - photo
              properties:
                photo:
                  type: string
                  description: Photo data
                  format: byte
                  minLength: 1
                  maxLength: 5242880
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  message:
                    type: string
                    description: Status
                  groupPhoto:
                    type: string
                    description: New photo
                    format: byte
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Username:
      type: object
      description: Username schema
      required:
        - name
      properties:
        name:
          type: string
          description: The username
          minLength: 3
          maxLength: 20
          pattern: "^[a-zA-Z0-9_]+$"
          example: "Alex"
    User:
      type: object
      description: User profile info
      required:
        - id
        - username
      properties:
        id:
          type: string
          description: User UUID
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
          example: "u1"
        username:
          $ref: "#/components/schemas/Username"
        profilePhoto:
          type: string
          description: Profile photo
          format: byte
          # nullable removed, use empty string or omission if needed
          minLength: 0
          maxLength: 5242880
    Reaction:
      type: object
      description: Message reaction
      required:
        - type
        - authorId
      properties:
        type:
          type: string
          description: Type of reaction
          enum: [like, heart, laugh, sad_face, angry_face]
        authorId:
          type: string
          description: User ID who reacted
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
          example: "u1"
    Message:
      type: object
      description: A message in a conversation
      required:
        - id
        - sender
        - content
        - timestamp
        - message_status
      properties:
        id:
          type: string
          description: Message UUID
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
          example: "msg1"
        sender:
          $ref: "#/components/schemas/User"
        content:
          type: object
          description: Message content
          required:
            - type
            - value
          properties:
            type:
              type: string
              description: Content type
              enum: [text, photo]
            value:
              type: string
              description: content value
              minLength: 1
              maxLength: 10000
        timestamp:
          type: string
          description: Message creation time
          format: date-time
          minLength: 20
          maxLength: 30
        message_status:
          type: string
          description: Status of the message
          enum: [sent, delivered, read]
        forwarded_from:
          type: string
          description: Original sender ID if forwarded
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
          # nullable: true
        reactions:
          type: array
          description: List of reactions
          minItems: 0
          maxItems: 100
          items:
            $ref: "#/components/schemas/Reaction"
          # nullable: true
    Conversation:
      type: object
      description: Conversation object
      required:
        - id
        - name
        - isGroup
      properties:
        id:
          type: string
          description: Conversation UUID
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
          example: "conv1"
        name:
          type: string
          description: Conversation name
          minLength: 0
          maxLength: 100
          # pattern: "^[a-zA-Z0-9 _\\-]*$"
          example: "Chat"
        groupPhoto:
          type: string
          description: Group photo
          format: byte
          minLength: 0
          maxLength: 5242880
          # nullable: true
        isGroup:
          type: boolean
          description: Is this a group conversation?
          example: true
        lastMessage:
          description: The last message in the conversation
          allOf:
            - $ref: "#/components/schemas/Message"
          # nullable: true
        participants:
          type: array
          description: List of participants
          minItems: 0
          maxItems: 100
          items:
            $ref: "#/components/schemas/User"
        messages:
          type: array
          description: List of messages
          minItems: 0
          maxItems: 1000
          items:
            $ref: "#/components/schemas/Message"
          # nullable: true